Component 의 기본 클래스 추가

+++ Chess/Core/Math.h
@@ -5,7 +5,7 @@ public:
     float y;
 
     Vector2() : x{ 0.f }, y{ 0.f }{};
-    Vector2(int inX, int inY) : x{ inX }, y{ inY }{};
+    Vector2(float inX, float inY) : x{ inX }, y{ inY }{};
 
     static const Vector2 Zero;
     static const Vector2 UnitX;
